python-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
vnstock==1.0.27
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2
pandas==2.1.4
numpy==1.24.3
asyncio-mqtt==0.16.1
pytest==7.4.3
pytest-asyncio==0.21.1
requests==2.31.0
